requirements.txt
tinyrpc==1.1.6
mysql-connector-python==9.0.0
python-dotenv==1.0.1
